mode save model settings to addAction
